src/requirements.txt
pymongo==4.6.1
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0

